move GET /api/scorer -> GET /api/scorer/nonce
